+++ packages/pangea-node-sdk/CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Retries now include a request header `X-Pangea-Retried-Request-Ids` to track
   the request IDs of the retries.
+- Simplified `PangeaResponse` export.
 
 ## 5.2.0 - 2025-07-21