+++ packages/pangea-node-sdk/CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## 5.3.0 - 2025-08-25
+
 ### Added
 
 - AI Guard: `transformed` field.